Textes d activité modifiables mais encore inutilement puisque visibles nulle part
