+++ app/application/config/tinyissue.php
@@ -1,6 +1,6 @@
 <?php
 return array(
-	'release_date' => '24-11-2021',
+	'release_date' => '25-11-2021',
 	'version' => '1.8',
-	'release' => '7ag'
+	'release' => '7ah'
 );
